removing redundant smart contracts
