+++ config/hardhat.config.json
@@ -7,6 +7,6 @@
   "cxipNFTDomainName": "https://nft.cxip.dev/",
   "ipfsDomainName": "https://ipfs.cxip.dev/",
   "arweaveDomainName": "https://arweave.cxip.dev/",
-  "identityProxyBytecode": "608060405234801561001057600080fd5b5060f68061001f6000396000f3fe6080604081905263071b938d60e31b8152600090735fbdb2315678afecb367f032d93f642f64180aa3906338dc9c6890608490602090600481865afa158015604b573d6000803e3d6000fd5b505050506040513d601f19601f82011682018060405250810190606d91906092565b90503660008037600080366000845af43d6000803e808015608d573d6000f35b3d6000fd5b60006020828403121560a357600080fd5b81516001600160a01b038116811460b957600080fd5b939250505056fea264697066735822122015f22c3ae06f0afaaabc521a9c187f7bef66f499b2667fdef2edcaa7ddb160ad64736f6c634300080c0033",
-  "erc721ProxyBytecode": "608060405234801561001057600080fd5b5060f68061001f6000396000f3fe60806040819052632c5feccb60e11b8152600090735fbdb2315678afecb367f032d93f642f64180aa3906358bfd99690608490602090600481865afa158015604b573d6000803e3d6000fd5b505050506040513d601f19601f82011682018060405250810190606d91906092565b90503660008037600080366000845af43d6000803e808015608d573d6000f35b3d6000fd5b60006020828403121560a357600080fd5b81516001600160a01b038116811460b957600080fd5b939250505056fea26469706673582212204996b86e77580c314351168858984744a572372b6231618c5b166e22158c56ef64736f6c634300080c0033"
+  "identityProxyBytecode": "608060405234801561001057600080fd5b5060f68061001f6000396000f3fe6080604081905263071b938d60e31b815260009073c267d41f81308d7773ecb3bdd863a902acc01ade906338dc9c6890608490602090600481865afa158015604b573d6000803e3d6000fd5b505050506040513d601f19601f82011682018060405250810190606d91906092565b90503660008037600080366000845af43d6000803e808015608d573d6000f35b3d6000fd5b60006020828403121560a357600080fd5b81516001600160a01b038116811460b957600080fd5b939250505056fea264697066735822122097155fb6bb02a768e72786a66b741a79a84ba8e2e116ada907f54065d9081f8864736f6c634300080c0033",
+  "erc721ProxyBytecode": "608060405234801561001057600080fd5b5060f68061001f6000396000f3fe60806040819052632c5feccb60e11b8152600090735fbdb2315678afecb367f032d93f642f64180aa3906358bfd99690608490602090600481865afa158015604b573d6000803e3d6000fd5b505050506040513d601f19601f82011682018060405250810190606d91906092565b90503660008037600080366000845af43d6000803e808015608d573d6000f35b3d6000fd5b60006020828403121560a357600080fd5b81516001600160a01b038116811460b957600080fd5b939250505056fea26469706673582212200ccd0771ef68a12b3c78ffcaf88afcf10e0d0f2a51e9296249fb5a9282c0b42664736f6c634300080c0033"
 }
