--- contracts/interface/ICxipCopyright.sol
@@ -1,17 +1,0 @@
-// SPDX-License-Identifier: MIT
-
-pragma solidity 0.8.12;
-
-/*______/\\\\\\\\\__/\\\_______/\\\__/\\\\\\\\\\\__/\\\\\\\\\\\\\___
- _____/\\\////////__\///\\\___/\\\/__\/////\\\///__\/\\\/////////\\\_
-  ___/\\\/_____________\///\\\\\\/________\/\\\_____\/\\\_______\/\\\_
-   __/\\\_________________\//\\\\__________\/\\\_____\/\\\\\\\\\\\\\/__
-    _\/\\\__________________\/\\\\__________\/\\\_____\/\\\/////////____
-     _\//\\\_________________/\\\\\\_________\/\\\_____\/\\\_____________
-      __\///\\\_____________/\\\////\\\_______\/\\\_____\/\\\_____________
-       ____\////\\\\\\\\\__/\\\/___\///\\\__/\\\\\\\\\\\_\/\\\_____________
-        _______\/////////__\///_______\///__\///////////__\///____________*/
-
-interface ICxipCopyright {
-
-}
