+++ contracts/proxy/CxipERC1155Proxy.sol
@@ -16,7 +16,7 @@ import "../interface/ICxipRegistry.sol";
 
 contract CxipERC1155Proxy {
     fallback() external payable {
-        address _target = ICxipRegistry(0xC267d41f81308D7773ecB3BDd863a902ACC01Ade)
+        address _target = ICxipRegistry(0x5FbDB2315678afecb367f032d93F642f64180aa3)
             .getERC1155CollectionSource();
         assembly {
             calldatacopy(0, 0, calldatasize())
