--- deploy/06_arsham.ts
@@ -1,24 +1,0 @@
-import { HardhatRuntimeEnvironment } from 'hardhat/types';
-import { DeployFunction } from 'hardhat-deploy/types';
-
-const func: DeployFunction = async function (hre: HardhatRuntimeEnvironment) {
-/*
-  const { deployments, getNamedAccounts } = hre;
-  const { deploy } = deployments;
-  const { deployer } = await getNamedAccounts();
-  const danielArshamErosions = await deploy('DanielArshamErodingAndReformingCars', {
-    from: deployer,
-    args: [],
-    log: true,
-  });
-
-  const danielArshamErosionsProxy = await deploy('DanielArshamErodingAndReformingCarsProxy', {
-    from: deployer,
-    args: [],
-    log: true,
-  });
-*/
-};
-export default func;
-func.tags = []; // ['DanielArshamErodingAndReformingCars', 'DanielArshamErodingAndReformingCarsProxy'];
-func.dependencies = ['CxipFactory', 'CxipRegistry'];
