--- deploy/07_identity.ts
@@ -1,17 +1,0 @@
-import { HardhatRuntimeEnvironment } from 'hardhat/types';
-import { DeployFunction } from 'hardhat-deploy/types';
-
-const func: DeployFunction = async function (hre: HardhatRuntimeEnvironment) {
-  const { deployments, getNamedAccounts } = hre;
-  const { deploy } = deployments;
-  const { deployer } = await getNamedAccounts();
-
-  const identity = await deploy('CxipIdentity', {
-    from: deployer,
-    args: [],
-    log: true,
-  });
-};
-export default func;
-func.tags = ['CxipIdentity'];
-func.dependencies = ['CxipFactory', 'CxipRegistry'];
