+++ deploy/08_verify.ts
@@ -5,19 +5,13 @@ const func: DeployFunction = async function (hre: HardhatRuntimeEnvironment) {
   const { deployments, getNamedAccounts } = hre;
 
   const cxipRegistry = await hre.ethers.getContract('CxipRegistry');
-  const cxipAssetProxy = await hre.ethers.getContract('CxipAssetProxy');
   const cxipERC721Proxy = await hre.ethers.getContract('CxipERC721Proxy');
-  const cxipIdentityProxy = await hre.ethers.getContract('CxipIdentityProxy');
   const cxipProvenanceProxy = await hre.ethers.getContract('CxipProvenanceProxy');
   const pA1DProxy = await hre.ethers.getContract('PA1DProxy');
-  const danielArshamErosionsProxy = await hre.ethers.getContract('DanielArshamErodingAndReformingCarsProxy');
 
   const cxipProvenance = await hre.ethers.getContract('CxipProvenance');
-  const cxipIdentity = await hre.ethers.getContract('CxipIdentity');
   const cxipERC721 = await hre.ethers.getContract('CxipERC721');
-  const cxipAsset = await hre.ethers.getContract('CxipAsset');
   const pA1D = await hre.ethers.getContract('PA1D');
-  const danielArshamErosions = await hre.ethers.getContract('DanielArshamErodingAndReformingCars');
 
   const mockErc721Receiver = await hre.ethers.getContract('MockERC721Receiver');
 
@@ -34,27 +28,11 @@ const func: DeployFunction = async function (hre: HardhatRuntimeEnvironment) {
   // Proxies
   try {
     await hre.run('verify:verify', {
-      address: cxipAssetProxy.address,
-      constructorArguments: [],
-    });
-  } catch (error) {
-    console.error(`Failed to verify AssetProxy ${error}`);
-  }
-  try {
-    await hre.run('verify:verify', {
       address: cxipERC721Proxy.address,
       constructorArguments: [],
     });
   } catch (error) {
     console.error(`Failed to verify ERC721Proxy ${error}`);
-  }
-  try {
-    await hre.run('verify:verify', {
-      address: cxipIdentityProxy.address,
-      constructorArguments: [],
-    });
-  } catch (error) {
-    console.error(`Failed to verify IdentityProxy ${error}`);
   }
   try {
     await hre.run('verify:verify', {
@@ -73,33 +51,7 @@ const func: DeployFunction = async function (hre: HardhatRuntimeEnvironment) {
     console.error(`Failed to verify PA1DProxy ${error}`);
   }
 
-  try {
-    await hre.run('verify:verify', {
-      address: pA1DProxy.address,
-      constructorArguments: [],
-    });
-  } catch (error) {
-    console.error(`Failed to verify PA1DProxy ${error}`);
-  }
-
-  try {
-    await hre.run('verify:verify', {
-      address: danielArshamErosionsProxy.address,
-      constructorArguments: [],
-    });
-  } catch (error) {
-    console.error(`Failed to verify DanielArshamErodingAndReformingCarsProxy ${error}`);
-  }
-
   // Implementations
-  try {
-    await hre.run('verify:verify', {
-      address: cxipAsset.address,
-      constructorArguments: [],
-    });
-  } catch (error) {
-    console.error(`Failed to verify Asset ${error}`);
-  }
   try {
     await hre.run('verify:verify', {
       address: cxipERC721.address,
@@ -107,14 +59,6 @@ const func: DeployFunction = async function (hre: HardhatRuntimeEnvironment) {
     });
   } catch (error) {
     console.error(`Failed to verify ERC721 ${error}`);
-  }
-  try {
-    await hre.run('verify:verify', {
-      address: cxipIdentity.address,
-      constructorArguments: [],
-    });
-  } catch (error) {
-    console.error(`Failed to verify Identity ${error}`);
   }
   try {
     await hre.run('verify:verify', {
@@ -131,14 +75,6 @@ const func: DeployFunction = async function (hre: HardhatRuntimeEnvironment) {
     });
   } catch (error) {
     console.error(`Failed to verify PA1D ${error}`);
-  }
-  try {
-    await hre.run('verify:verify', {
-      address: danielArshamErosions.address,
-      constructorArguments: [],
-    });
-  } catch (error) {
-    console.error(`Failed to verify DanielArshamErodingAndReformingCars ${error}`);
   }
   try {
     await hre.run('verify:verify', {
