--- src/proxy/DanielArshamErodingAndReformingCarsProxy.sol
@@ -1,72 +1,0 @@
-// SPDX-License-Identifier: MIT
-
-pragma solidity 0.8.12;
-
-/*
-
-    ██████╗  █████╗ ███╗   ██╗██╗███████╗██╗
-    ██╔══██╗██╔══██╗████╗  ██║██║██╔════╝██║
-    ██║  ██║███████║██╔██╗ ██║██║█████╗  ██║
-    ██║  ██║██╔══██║██║╚██╗██║██║██╔══╝  ██║
-    ██████╔╝██║  ██║██║ ╚████║██║███████╗███████╗
-    ╚═════╝ ╚═╝  ╚═╝╚═╝  ╚═══╝╚═╝╚══════╝╚══════╝
-
-  █████╗ ██████╗ ███████╗██╗  ██╗ █████╗ ███╗   ███╗
- ██╔══██╗██╔══██╗██╔════╝██║  ██║██╔══██╗████╗ ████║
- ███████║██████╔╝███████╗███████║███████║██╔████╔██║
- ██╔══██║██╔══██╗╚════██║██╔══██║██╔══██║██║╚██╔╝██║
- ██║  ██║██║  ██║███████║██║  ██║██║  ██║██║ ╚═╝ ██║
- ╚═╝  ╚═╝╚═╝  ╚═╝╚══════╝╚═╝  ╚═╝╚═╝  ╚═╝╚═╝     ╚═╝
-
-                       ______
-                      /     /\
-                     /     /##\
-                    /     /####\
-                   /     /######\
-                  /     /########\
-                 /     /##########\
-                /     /#####/\#####\
-               /     /#####/++\#####\
-              /     /#####/++++\#####\
-             /     /#####/\+++++\#####\
-            /     /#####/  \+++++\#####\
-           /     /#####/    \+++++\#####\
-          /     /#####/      \+++++\#####\
-         /     /#####/        \+++++\#####\
-        /     /#####/__________\+++++\#####\
-       /                        \+++++\#####\
-      /__________________________\+++++\####/
-      \+++++++++++++++++++++++++++++++++\##/
-       \+++++++++++++++++++++++++++++++++\/
-        ``````````````````````````````````
-
-              ██████╗██╗  ██╗██╗██████╗
-             ██╔════╝╚██╗██╔╝██║██╔══██╗
-             ██║      ╚███╔╝ ██║██████╔╝
-             ██║      ██╔██╗ ██║██╔═══╝
-             ╚██████╗██╔╝ ██╗██║██║
-              ╚═════╝╚═╝  ╚═╝╚═╝╚═╝
-
-*/
-
-import "../interface/ICxipRegistry.sol";
-
-// sha256(abi.encodePacked('eip1967.CxipRegistry.DanielArshamErodingAndReformingCarsProxy')) == 0xa02fc078e74005974d5615d21c608de70bf6b5bb5d4859bca6aeb16e41be6ff9
-contract DanielArshamErodingAndReformingCarsProxy {
-    fallback() external payable {
-        // sha256(abi.encodePacked('eip1967.CxipRegistry.DanielArshamErodingAndReformingCars')) == 0xe3b4c4e0b41f8dc247603a686e2acd61e0a5b5d2a95ce2e35a1744406075c82f
-        address _target = ICxipRegistry(0xdeaDDeADDEaDdeaDdEAddEADDEAdDeadDEADDEaD).getCustomSource(0xe3b4c4e0b41f8dc247603a686e2acd61e0a5b5d2a95ce2e35a1744406075c82f);
-        assembly {
-            calldatacopy(0, 0, calldatasize())
-            let result := delegatecall(gas(), _target, 0, calldatasize(), 0, 0)
-            returndatacopy(0, 0, returndatasize())
-            switch result
-            case 0 {
-                revert(0, returndatasize())
-            }
-            default {
-                return(0, returndatasize())
-            }
-        }
-    }
-}
